+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/QuestList.kt
@@ -373,7 +373,7 @@ class QuestList : Fragment(R.layout.quest_list) {
         viewModel.message.observe(viewLifecycleOwner) { message ->
             message.getContentIfNotHandled()?.let {
                 val addNewQuestButton: View = view.findViewById(R.id.AddNewQuestButton)
-                showSnackbar(it, requireView(), addNewQuestButton)
+                showSnackbar(it, view, addNewQuestButton)
             }
         }
     }